% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_getDR.R
\name{cc_getDR}
\alias{cc_getDR}
\title{Generate Dose-Response Data and Plots}
\usage{
cc_getDR(norm_data, name_char_limit = 16, rounding_value_IC50 = 2)
}
\arguments{
\item{norm_data}{List. Output from `cc_norm`}

\item{name_char_limit}{Numeric. Maximum number of characters to display in the drug name on plots. Default is `16`.}

\item{rounding_value_IC50}{Numeric. Number of decimal places to round IC50 values in the plot titles. Default is `2`.}
}
\description{
This function processes normalized data to fit dose-response models, calculate IC50 values, 
and generate dose-response plots for both dense and sparse mode data.
}
