% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_makeMeta.R
\name{cc_makeMeta}
\alias{cc_makeMeta}
\title{Generate 'Complete' Metadata for Combocat Experiments}
\usage{
cc_makeMeta(
  reg_file,
  data_mode = "dense",
  dilution_factor = 3,
  assay_ready_volume = "auto_detect",
  final_volume = "standard",
  sample_name = "not_provided",
  rounding_digits = 6
)
}
\arguments{
\item{reg_file}{Data frame. A file containing plate registration and Echo transfer logs (`.csv` format).}

\item{data_mode}{Character. Specifies the data mode: `"dense"` or `"sparse"`.}

\item{dilution_factor}{Numeric. The dilution factor used for all drugs. Default is `3` (i.e., 1:3 dilution). Multiple dilution factors are not supported for this function.}

\item{assay_ready_volume}{Character or Numeric. The total expected volume of the assay-ready plate in nanoliters (nL). Defaults to `"auto_detect"`, which uses the most frequent volume in the `Amount` column. In sparse mode, the standard volume is 20 nL. In dense mode, the standard volume is 200nL.}

\item{final_volume}{Character or Numeric. The final volume of the assay in NANOliters. Default is `"standard"` which is 40,000nL (40uL) in dense mode or 4,000nL (4uL) sparse mode.}

\item{sample_name}{Character. The name of the sample. Defaults to `"not_provided"`.}
}
\description{
This function generates "complete" type metadata for a Combocat experiment (either sparse or dense mode). 
It requires a very specific registration file containing plate registration and Echo transfer logs. The raw data files must be named as `[Plate Barcode].csv` where `[Plate Barcode]` is the corresponding column in the registration file.
}
