% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_plotExtras.R
\name{cc_plotExtras}
\alias{cc_plotExtras}
\title{Generate Supplemental Plots for Synergy Analysis}
\usage{
cc_plotExtras(
  syn_data,
  color_midpoint = 20,
  rounding_value = 4,
  name_char_limit = 16
)
}
\arguments{
\item{syn_data}{List. Output from the `cc_getSyn`}

\item{color_midpoint}{Numeric. Midpoint for the color scale in scatterplots. Default is `20`.}

\item{rounding_value}{Numeric. Number of decimal places to round the drug concentrations for plotting. Default is `4`.}

\item{name_char_limit}{Numeric. Maximum number of characters to use for drug names in plots. Default is `16`.}
}
\description{
This function creates supplemental plots for drug combination synergy analysis, including barplots of %Cell Death for maximum synergy and scatterplots of observed vs. expected cell death.
}
