% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_predict.R
\name{cc_predict}
\alias{cc_predict}
\title{Predict Cell Death Values for Sparse Matrices}
\usage{
cc_predict(models_list, norm_data)
}
\arguments{
\item{models_list}{List. The `models_list.RDS` file output from `cc_buildModel`, containing serialized machine learning models.}

\item{norm_data}{List. Output from `cc_norm`, containing normalized data for sparse matrices.}
}
\description{
This function predicts 90 cell death values for sparse matrices using a machine learning model. 
It updates the input data (`norm_data`) by adding predicted values to the reference data frame (`ref_df`) 
and populates the sparse matrix (`mean_perc_cell_death_mat`) with predicted values.
}
