% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_report.R
\name{cc_report}
\alias{cc_report}
\title{Generate a Summary of All Results}
\usage{
cc_report(
  syn_data,
  dr_data,
  cd_plots,
  syn_plots,
  extra_plots,
  save_summary_files = TRUE,
  save_summary_plots = TRUE
)
}
\arguments{
\item{syn_data}{List. Output from `cc_getSyn`}

\item{dr_data}{List. Output from `cc_getDr`}

\item{cd_plots}{List. Output from `cc_plotMat`, using cell death as input.}

\item{syn_plots}{List. Output from `cc_plotMat`, using synergy as input.}

\item{extra_plots}{List. Output from `cc_plotExtras`}

\item{save_summary_files}{Logical. Whether to save combined `ref_df` and `IC50_df` data frames as CSV files. Defaults to `TRUE`.}

\item{save_summary_plots}{Logical. Whether to save summary plots as `.svg` files. Defaults to `TRUE`.}
}
\description{
This function combines results from multiple steps (synergy, dose-response, and plots)
into a comprehensive summary. It saves summary files and plots as specified by the user.
}
